Prezzi da float a double per precisione corretta

+++ PromoCalculator/BaseTypes.h
@@ -11,7 +11,7 @@
 
 struct Totals {
     long itemsNumber ;
-    float totalAmount ;
+    double totalAmount ;
 };
 
 struct CartRow {
